Reset Consumes annotation to accept everything for (un)schedule (#8981)
    Since these endpoints don't accept a request body, we can just accept
    any content-type. This unbreaks calling these endpoints with our swagger
    API browser.

+++ dataset/GitHub_Java/Graylog2.graylog2-server/graylog2-server/src/main/java/org/graylog/events/rest/EventDefinitionsResource.java
@@ -22,7 +22,9 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
+import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.graylog.events.audit.EventsAuditEventTypes;
+import org.graylog.events.context.EventDefinitionContextService;
 import org.graylog.events.processor.DBEventDefinitionService;
 import org.graylog.events.processor.EventDefinitionDto;
 import org.graylog.events.processor.EventDefinitionHandler;
@@ -32,6 +34,7 @@ import org.graylog.events.processor.EventProcessorParameters;
 import org.graylog.events.processor.EventProcessorParametersWithTimerange;
 import org.graylog2.audit.jersey.AuditEvent;
 import org.graylog2.audit.jersey.NoAuditEvent;
+import org.graylog2.database.PaginatedList;
 import org.graylog2.plugin.rest.PluginRestResource;
 import org.graylog2.plugin.rest.ValidationResult;
 import org.graylog2.rest.models.PaginatedResponse;
@@ -39,6 +42,9 @@ import org.graylog2.search.SearchQuery;
 import org.graylog2.search.SearchQueryField;
 import org.graylog2.search.SearchQueryParser;
 import org.graylog2.shared.rest.resources.RestResource;
+import org.graylog2.shared.security.RestPermissions;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
 import javax.validation.Valid;
@@ -48,6 +54,7 @@ import javax.ws.rs.BadRequestException;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.DefaultValue;
+import javax.ws.rs.ForbiddenException;
 import javax.ws.rs.GET;
 import javax.ws.rs.InternalServerErrorException;
 import javax.ws.rs.NotFoundException;
@@ -59,6 +66,9 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Api(value = "Events/Definitions", description = "Event definition management")
 @Path("/events/definitions")
@@ -66,6 +76,8 @@ import javax.ws.rs.core.Response;
 @Consumes(MediaType.APPLICATION_JSON)
 @RequiresAuthentication
 public class EventDefinitionsResource extends RestResource implements PluginRestResource {
+    private static final Logger LOG = LoggerFactory.getLogger(EventDefinitionsResource.class);
+
     private static final ImmutableMap<String, SearchQueryField> SEARCH_FIELD_MAPPING = ImmutableMap.<String, SearchQueryField>builder()
             .put("id", SearchQueryField.create(EventDefinitionDto.FIELD_ID))
             .put("title", SearchQueryField.create(EventDefinitionDto.FIELD_TITLE))
@@ -74,15 +86,18 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
 
     private final DBEventDefinitionService dbService;
     private final EventDefinitionHandler eventDefinitionHandler;
+    private final EventDefinitionContextService contextService;
     private final EventProcessorEngine engine;
     private final SearchQueryParser searchQueryParser;
 
     @Inject
     public EventDefinitionsResource(DBEventDefinitionService dbService,
                                     EventDefinitionHandler eventDefinitionHandler,
+                                    EventDefinitionContextService contextService,
                                     EventProcessorEngine engine) {
         this.dbService = dbService;
         this.eventDefinitionHandler = eventDefinitionHandler;
+        this.contextService = contextService;
         this.engine = engine;
         this.searchQueryParser = new SearchQueryParser(EventDefinitionDto.FIELD_TITLE, SEARCH_FIELD_MAPPING);
     }
@@ -93,14 +108,32 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
                                                       @ApiParam(name = "per_page") @QueryParam("per_page") @DefaultValue("50") int perPage,
                                                       @ApiParam(name = "query") @QueryParam("query") @DefaultValue("") String query) {
         final SearchQuery searchQuery = searchQueryParser.parse(query);
-        return PaginatedResponse.create("event_definitions", dbService.getAllPaginated(searchQuery, "title", page, perPage), query);
+        final PaginatedList<EventDefinitionDto> result = dbService.searchPaginated(searchQuery, event -> {
+            return isPermitted(RestPermissions.EVENT_DEFINITIONS_READ, event.id());
+        }, "title", page, perPage);
+        final ImmutableMap<String, Object> context = contextService.contextFor(result.delegate());
+        return PaginatedResponse.create("event_definitions", result, query, context);
     }
 
     @GET
     @Path("{definitionId}")
     @ApiOperation("Get an event definition")
     public EventDefinitionDto get(@ApiParam(name = "definitionId") @PathParam("definitionId") @NotBlank String definitionId) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_READ, definitionId);
         return dbService.get(definitionId)
+                .orElseThrow(() -> new NotFoundException("Event definition <" + definitionId + "> doesn't exist"));
+    }
+
+    @GET
+    @Path("{definitionId}/with-context")
+    @ApiOperation("Get an event definition")
+    public Map<String, Object> getWithContext(@ApiParam(name = "definitionId") @PathParam("definitionId") @NotBlank String definitionId) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_READ, definitionId);
+        return dbService.get(definitionId)
+                .map(evenDefinition -> ImmutableMap.of(
+                        "event_definition", evenDefinition,
+                        "context", contextService.contextFor(evenDefinition)
+                ))
                 .orElseThrow(() -> new NotFoundException("Event definition <" + definitionId + "> doesn't exist"));
     }
 
@@ -108,12 +141,17 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     @Produces(MediaType.APPLICATION_JSON)
     @ApiOperation("Create new event definition")
     @AuditEvent(type = EventsAuditEventTypes.EVENT_DEFINITION_CREATE)
-    public Response create(EventDefinitionDto dto) {
-        final ValidationResult result = dto.config().validate();
+    @RequiresPermissions(RestPermissions.EVENT_DEFINITIONS_CREATE)
+    public Response create(@ApiParam("schedule") @QueryParam("schedule") @DefaultValue("true") boolean schedule,
+                           @ApiParam(name = "JSON Body") EventDefinitionDto dto) {
+        checkEventDefinitionPermissions(dto, "create");
+
+        final ValidationResult result = dto.validate();
         if (result.failed()) {
             return Response.status(Response.Status.BAD_REQUEST).entity(result).build();
         }
-        return Response.ok().entity(eventDefinitionHandler.create(dto)).build();
+        final EventDefinitionDto entity = schedule ? eventDefinitionHandler.create(dto) : eventDefinitionHandler.createWithoutSchedule(dto);
+        return Response.ok().entity(entity).build();
     }
 
     @PUT
@@ -121,11 +159,14 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     @ApiOperation("Update existing event definition")
     @AuditEvent(type = EventsAuditEventTypes.EVENT_DEFINITION_UPDATE)
     public Response update(@ApiParam(name = "definitionId") @PathParam("definitionId") @NotBlank String definitionId,
-                           EventDefinitionDto dto) {
+                           @ApiParam("schedule") @QueryParam("schedule") @DefaultValue("true") boolean schedule,
+                           @ApiParam(name = "JSON Body") EventDefinitionDto dto) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_EDIT, definitionId);
+        checkEventDefinitionPermissions(dto, "update");
         dbService.get(definitionId)
                 .orElseThrow(() -> new NotFoundException("Event definition <" + definitionId + "> doesn't exist"));
 
-        final ValidationResult result = dto.config().validate();
+        final ValidationResult result = dto.validate();
         if (!definitionId.equals(dto.id())) {
             result.addError("id", "Event definition IDs don't match");
         }
@@ -133,7 +174,7 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
         if (result.failed()) {
             return Response.status(Response.Status.BAD_REQUEST).entity(result).build();
         }
-        return Response.ok().entity(eventDefinitionHandler.update(dto)).build();
+        return Response.ok().entity(eventDefinitionHandler.update(dto, schedule)).build();
     }
 
     @DELETE
@@ -141,7 +182,28 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     @ApiOperation("Delete event definition")
     @AuditEvent(type = EventsAuditEventTypes.EVENT_DEFINITION_DELETE)
     public void delete(@ApiParam(name = "definitionId") @PathParam("definitionId") @NotBlank String definitionId) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_DELETE, definitionId);
         eventDefinitionHandler.delete(definitionId);
+    }
+
+    @PUT
+    @Path("{definitionId}/schedule")
+    @Consumes(MediaType.WILDCARD)
+    @ApiOperation("Enable event definition")
+    @AuditEvent(type = EventsAuditEventTypes.EVENT_DEFINITION_UPDATE)
+    public void schedule(@ApiParam(name = "definitionId") @PathParam("definitionId") @NotBlank String definitionId) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_EDIT, definitionId);
+        eventDefinitionHandler.schedule(definitionId);
+    }
+
+    @PUT
+    @Path("{definitionId}/unschedule")
+    @Consumes(MediaType.WILDCARD)
+    @ApiOperation("Disable event definition")
+    @AuditEvent(type = EventsAuditEventTypes.EVENT_DEFINITION_UPDATE)
+    public void unschedule(@ApiParam(name = "definitionId") @PathParam("definitionId") @NotBlank String definitionId) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_EDIT, definitionId);
+        eventDefinitionHandler.unschedule(definitionId);
     }
 
     @POST
@@ -150,6 +212,7 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     @AuditEvent(type = EventsAuditEventTypes.EVENT_DEFINITION_EXECUTE)
     public void execute(@ApiParam(name = "definitionId") @PathParam("definitionId") @NotBlank String definitionId,
                         @ApiParam(name = "parameters", required = true) @NotNull EventProcessorParameters parameters) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_EXECUTE, definitionId);
         if (parameters instanceof EventProcessorParametersWithTimerange.FallbackParameters) {
             throw new BadRequestException("Unknown parameters type");
         }
@@ -165,8 +228,20 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     @Path("/validate")
     @NoAuditEvent("Validation only")
     @ApiOperation(value = "Validate an event definition")
+    @RequiresPermissions(RestPermissions.EVENT_DEFINITIONS_CREATE)
     public ValidationResult validate(@ApiParam(name = "JSON body", required = true)
                                      @Valid @NotNull EventDefinitionDto toValidate) {
         return toValidate.config().validate();
+    }
+
+    private void checkEventDefinitionPermissions(EventDefinitionDto dto, String action) {
+        final Set<String> missingPermissions = dto.requiredPermissions().stream()
+                .filter(permission -> !isPermitted(permission))
+                .collect(Collectors.toSet());
+
+        if (!missingPermissions.isEmpty()) {
+            LOG.info("Not authorized to {} event definition. User <{}> is missing permissions: {}", action, getSubject().getPrincipal(), missingPermissions);
+            throw new ForbiddenException("Not authorized");
+        }
     }
 }
